refactor(helpers): some functions moved from MyWallet to Helpers.
