+++ src/address.js
@@ -181,7 +181,7 @@ Address.fromString = function(keyOrAddr, label, bipPass){
       return resolve(Address.import(keyOrAddr, label));
     } else {
       // Import private key
-      var format    = MyWallet.detectPrivateKeyFormat(keyOrAddr)
+      var format    = Helpers.detectPrivateKeyFormat(keyOrAddr)
         , okFormats = ['base58', 'base64', 'hex', 'mini', 'sipa', 'compsipa'];
 
       if (format === 'bip38') {
